feat: ✨ mv command replaced with js, extracted few helpers
