+++ package.json
@@ -38,8 +38,8 @@
     "build": "npm run build:tsc && rollup -c config/rollup.config.js && rollup -c config/rollup.config.js --environment NODE_ENV:production",
     "build:tsc": "tsc -p ./config/tsconfig.build.json && tsc -p ./config/tsconfig.build.json --target es2018 --outDir dist/esm2015",
     "postbuild": "npm run rename-mjs && npm run size",
-    "rename-mjs": "mv dist/bundles/index.esm.js dist/bundles/index.mjs && mv dist/bundles/index.esm.min.js dist/bundles/index.min.mjs ",
-    "size": "node config/file-size ./dist/bundles/index.min.mjs ./dist/bundles/index.umd.min.js",
+    "rename-mjs": "node ./config/renameMjs.js",
+    "size": "node config/fileSize ./dist/bundles/index.min.mjs ./dist/bundles/index.umd.min.js",
     "serve": "ws --directory dev/www --port 3000 --spa index.html",
     "clean": "rimraf 'node_modules/.cache' '.awcache' 'dist'",
     "pull-changes": "git checkout master && git pull && git checkout develop && git pull",
